Add files via upload
 Updated 6 minutes ago
The project aims to control an autonomous vehicle, approximated by a bicycle model, to track a reference trajectory in R^2. It is assumed that the track is known T seconds in advance. Current techniques for tracking involve Model predictive Control, with a high time complexity. This control algorithm solves the same problem by predicting future state using forward Euler on current system, keeping the input constant. The error between the predicted position and the required position is used to change the system input.

Run the main file to execute the project. The study is done for 3 speeds- 15km/hr, 25km/hr, 35km/hr